refactor: Move blockstore interface close to extern
