+++ wnfs-wasm/src/fs/blockstore.rs
@@ -13,6 +13,15 @@ use wnfs::common::BlockStore as WnfsBlockStore;
 // Externs
 //--------------------------------------------------------------------------------------------------
 
+#[wasm_bindgen(typescript_custom_section)]
+const TS_BLOCKSTORE: &'static str = r#"
+export interface BlockStore {
+    putBlockKeyed(cid: Uint8Array, bytes: Uint8Array): Promise<void>;
+    getBlock(cid: Uint8Array): Promise<Uint8Array | undefined>;
+    hasBlock(cid: Uint8Array): Promise<boolean>;
+}
+"#;
+
 #[wasm_bindgen]
 extern "C" {
     #[wasm_bindgen(typescript_type = "BlockStore")]
